Fix comments.
